+++ keyconnect-cli/src/main/java/app/keyconnect/cli/commands/PayCommand.java
@@ -6,8 +6,8 @@ import app.keyconnect.cli.utils.ConsoleUtil;
 import app.keyconnect.cli.utils.LocalWalletData;
 import app.keyconnect.cli.utils.LocalWalletHelper;
 import app.keyconnect.sdk.wallets.BlockchainWallet;
-import app.keyconnect.sdk.wallets.Payment;
-import app.keyconnect.sdk.wallets.SubmittedPayment;
+import app.keyconnect.sdk.payments.Payment;
+import app.keyconnect.sdk.payments.SubmittedPayment;
 import java.math.BigDecimal;
 import java.util.Locale;
 import java.util.Optional;
