+++ keyconnect-sdk/src/main/java/app/keyconnect/sdk/wallets/BlockchainWallet.java
@@ -19,6 +19,7 @@ public interface BlockchainWallet {
 
   /**
    * User-friendly name of the wallet
+   *
    * @return Wallet name
    */
   String getName();
